requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.9.4
pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
django-bootstrap4==0.0.8
pyserial==3.4